refactor: split renderer draw module
